Added @Authorized annotations to service methods
AppointmentQueryService.java:
Added authorized annotations to service methods to ensure you are logged
in before accessing services. Privilege values may be added to the
annotations as permissions are set up and configured.
AppointmentDataService.java:
Added authorized annotations to service methods to ensure you are logged
it before accessing services. Privilege values may be added to the
annotations as permissions are set up and configured.